--- Spurning 2 fraedi
@@ -1,53 +1,0 @@
-# *Spurning 2: Leita af netfangi*
-
-Reglulega segðin:
-```regex
-r'^[a-zA-Z0-9_.+-]+@[a-zA-Z0-9-]+\.[a-zA-Z]{2,3}(?:\.[a-zA-Z]{2,})?$'
-```
-
-Þessi reglulega segð leitar af leyfilegu netfangi:
-```regex
-^ 
-```
-Byrjun á streng.
-
-
-```regex
-[a-zA-Z0-9_.+-]+ 
-```
-
-Passar notendanafnið (bókstafir, tölustafir og sérstafir eins og punktur, undirstika o.s.frv.).
-```regex
-@ 
-```
-Táknar að það sé „at“ merki.
-
-```regex
-[a-zA-Z0-9-]+
-```
-Passar  lén (bókstafir og tölustafir, með bandstrik mögulegt).```
-
-```regex
-\. 
-```
-Táknar punkt á undan lénaviðbót.
-
-
-```regex
-[a-zA-Z]{2,3}
-```
-Passar á tveggja eða þriggja stafa lénaviðbót (t.d. .com, .is).
-```
-(?:\.[a-zA-Z]{2,})?
-```
-Valfrjáls lengri viðbót fyrir lén (t.d. .co.uk).
-```regex
-$ 
-```
-lok strengs.
-
-
-Mögulega þarf að nota þennan kóða til að keyra file fyrir Spurningu 2
-```regex
-python3 sp2.py --file data/email.txt
-```
