+++ code/regex_reorder.py
@@ -1,17 +1,16 @@
 import os
 import argparse
-
+import re
 
 def parse_arguments():
     """
     Lesa inn argument frá skipanalínu
     :return: (ArgumentParser) parser með argumentunum
     """
-
     parser = argparse.ArgumentParser(description="""Endurraða csv-skrá sem hefur röðunina: 
     1) eiginnafn millinafn kenninafn, 2) heimilisfang, 3) símanúmer.    
     Úttaksskráin verður endurröðuð í tsv röð:
-    1) heimilisfang, 2) símanúmer, 3) kenninafn, eigninnafn millinafn.
+    1) heimilisfang, 2) símanúmer, 3) kenninafn, eiginnafn millinafn.
     """)
     parser.add_argument('--infile', required=True,
                         help='Slóð að inntaksskrá með upplýsingum.')
@@ -26,17 +25,26 @@ def lesa_skra(file_path):
     :param file_path: (str) Slóð að skrá
     :return:          (list) Listi af línum
     """
-
     with open(file_path, 'r') as file:
         return file.readlines()
 
 
 def endurraða_skra(linur):
     """"
-    Hér á eftir að uppfæra doc-streng sem lýsir fallinu betur.
+    Endurröðun lína úr CSV formi í TSV form með notkun reglulegra segða.
+    :param linur: (list) Listi af línum úr CSV skránni
+    :return:      (list) Listi af endurröðuðum línum
     """
+    
+    nyjar_linur = []
+    # Regluleg segð til að finna nafn, heimilisfang, póstnúmer og símanúmer
+    segd = r"^(\w+(?:\s+\w+)?)\s+(\w+),\s*(.*?),\s*(\d{3}.*?),\s(.*?)$"
 
-    raise NotImplementedError("Regluleg segð til að endurraða línur hefur ekki verið útfærð.")
+    for lina in linur:
+        ny_lina = re.sub(segd, r"\3\t\4\t\5\t\2, \1", lina.strip())
+        nyjar_linur.append(ny_lina)
+    
+    return nyjar_linur
 
 
 def skrifa_nidurstodur(output_file, linur):
@@ -83,3 +91,4 @@ def main():
 
 if __name__ == "__main__":
     main()
+
