--- sp3.py
@@ -1,93 +1,0 @@
-import os
-import argparse
-import re
-
-def parse_arguments():
-    """
-    Lesa inn argument frá skipanalínu
-    :return: (ArgumentParser) parser með argumentunum
-    """
-    parser = argparse.ArgumentParser(description="""Endurraða csv-skrá sem hefur röðunina: 
-    1) eiginnafn millinafn kenninafn, 2) heimilisfang, 3) símanúmer.    
-    Úttaksskráin verður endurröðuð í tsv röð:
-    1) heimilisfang, 2) símanúmer, 3) kenninafn, eiginnafn millinafn.
-    """)
-    parser.add_argument('--infile', required=True,
-                        help='Slóð að inntaksskrá með upplýsingum.')
-    parser.add_argument('--outfile', required=True,
-                        help='Slóð að úttaksskrá með endurröðuðum upplýsingum.')
-    return parser.parse_args()
-
-
-def lesa_skra(file_path):
-    """
-    Lesa inntaksskrá og skila lista af línum
-    :param file_path: (str) Slóð að skrá
-    :return:          (list) Listi af línum
-    """
-    with open(file_path, 'r') as file:
-        return file.readlines()
-
-
-def endurraða_skra(linur):
-    """"
-    Endurröðun lína úr CSV formi í TSV form með notkun reglulegra segða.
-    :param linur: (list) Listi af línum úr CSV skránni
-    :return:      (list) Listi af endurröðuðum línum
-    """
-    
-    nyjar_linur = []
-    # Regluleg segð til að finna nafn, heimilisfang, póstnúmer og símanúmer
-    segd = r"^(\w+(?:\s+\w+)?)\s+(\w+),\s*(.*?),\s*(\d{3}.*?),\s(.*?)$"
-
-    for lina in linur:
-        ny_lina = re.sub(segd, r"\3\t\4\t\5\t\2, \1", lina.strip())
-        nyjar_linur.append(ny_lina)
-    
-    return nyjar_linur
-
-
-def skrifa_nidurstodur(output_file, linur):
-    """
-    Skrifar niðurstöður í úttaksskrá
-    :param output_file: (str) Slóð að úttaksskrá
-    :param linur:       (list) Listi af línum
-    :return:            None
-    """
-    with open(output_file, 'w') as file:
-        for lina in linur:
-            file.write(lina + '\n')
-
-
-def main():
-    """
-    Keyrslufall sem les inn skrá, endurraðar henni og skrifar niðurstöður í nýja skrá.
-    :return: None
-    """
-
-    args = parse_arguments()
-
-    # Athuga hvort skráar endingar séu réttar
-    if not args.infile.endswith('.csv'):
-        print('Inntaksskrá þarf að vera csv skrá.')
-        return
-    if not args.outfile.endswith('.tsv'):
-        print('Úttaksskrá þarf að vera tsv skrá.')
-        return
-    # Athuga hvort inntaksskrá sé til
-    if not os.path.exists(args.infile):
-        print('Inntaksskrá fannst ekki.')
-        return
-
-    # Lesa línur úr skránni
-    linur = lesa_skra(args.infile)
-
-    # Endurraða hverri línu í skránni
-    linur = endurraða_skra(linur)
-
-    # Skrifa niðurstöðurnar í nýja skrá
-    skrifa_nidurstodur(args.outfile, linur)
-
-
-if __name__ == "__main__":
-    main()
